Fix memory leak when stopping camera (#126)
* Clear the buffer and resets texture and texture registry on method call stop()
* Moves the texture/buffer handling to QrReader
