+++ ios/Classes/SwiftQrMobileVisionPlugin.swift
@@ -62,13 +62,11 @@ public class SwiftQrMobileVisionPlugin: NSObject, FlutterPlugin {
       }
 
       let options = VisionBarcodeDetectorOptions(formatStrings: formatStrings)
-      
-      let texture = TextureHandler(registry: textureRegistry)
-      
+            
       reader = QrReader(
         targetWidth: targetWidth,
         targetHeight: targetHeight,
-        textureHandler: texture,
+        textureRegistry: textureRegistry,
         options: options) { [unowned self] qr in
           self.channel.invokeMethod("qrRead", arguments: qr)
       }
@@ -79,7 +77,7 @@ public class SwiftQrMobileVisionPlugin: NSObject, FlutterPlugin {
         "surfaceWidth": reader!.previewSize.height,
         "surfaceHeight": reader!.previewSize.width,
         "surfaceOrientation": 0, //TODO: check on iPAD
-        "textureId": texture.textureId!
+        "textureId": reader!.textureId!
       ])
     case "stop":
       reader?.stop();
